[fix] delete types (they in stdint), update interupts declaration (cli and sti), delete extra flags in makefile, change r to *r in registers_t

+++ lib/mem.c
@@ -1,4 +1,5 @@
 #include "mem.h"
+
 void memory_copy(uint8_t* source, uint8_t* dest, size_t nbytes) {
 	for (size_t i = 0; i < nbytes; ++i) {
 		*(dest + i) = *(source + i);
@@ -14,7 +15,7 @@ void memory_set(uint8_t* dest, uint8_t val, uint32_t len) {
 
 uint32_t free_memory_addr = 0x10000;
 
-uint32_t kernel_malloc(uint32_t size, int align, uint32_t* phys_addr) {
+uint32_t kernel_malloc(size_t size, int align, uint32_t* phys_addr) {
 	if (align == 1 && (free_memory_addr & 0xFFF)) {
 		free_memory_addr &= 0xFFFFF000;
 		free_memory_addr += 0x1000;
